// backend/src/api/v1/mentorShip/services/types.ts
export interface userData {
    name: string
    email: string
    dateOfBirth: string
    yearsOfExperience: number
    currentJobTitle: string
    areasOfExpertise: string[]
    rate: number
    interests: string[]
    aboutMe: string
    id: string
    isMentor: boolean
}

export interface articleData {
    articleName: string
    url: string
}